Updating Scripts
Created the singleton "Parts" script to keep a list of parts and their
bonuses so that the race scene and the car customize scene can both
benefit from it.

I also changed a little in the "CarMods" script to work with the new
"Parts" script and fixed the errors paralyzing the project.